Allowed for registering of nodes on other nodes.
Previously agents could register direct connections to eachother but not
routes to eachother.

Now they can, as registration is maintained by agentRegisterer.

There's also a bunch of methods added to various classes to work with
this. MetaAgent needs revising because of this, as there may be an
increase in requirements for an agent.