Removed routing fallback and added empty JUnit tests
--Removed routing fallback --

Before there was a fallback for sending messages where a portal would
check if any attached portals had knowledge of the agent. This made the
system more likely to succeed but meant that the rules of the routing
were completely overlooked - in an ideal system a portal should only be
able to find an agent if it's registered with it. This was originally
implemented to ensure deadlines were met but has now been removed and
instead it is required to register it.

--added empty JUnit tests--

JUnit tests have been generated using netbeans and are in the process of
being written (With one completed already). Hopefully they all pass and
will shine some light on any issues present in the system.